# docs/swagger.yaml
basePath: /
definitions:
  common.AdvertFieldParams:
    properties:
      fields:
        $ref: '#/definitions/common.Fields'
    type: object
  common.AdvertSortOrderParams:
    properties:
      created_at:
        description: по дате создания
        type: integer
      offset:
        description: смещение от 0
        type: integer
      price:
        description: сортировка по цене
        type: integer
    type: object
  common.AdvertWithPhoto:
    properties:
      description:
        type: string
      photos:
        description: массив фото
        items:
          $ref: '#/definitions/common.Photo'
        type: array
      price:
        type: number
      title:
        type: string
    type: object
  common.Fields:
    properties:
      all_photo:
        description: включить в ответ все фото, а не только главное, необязательное
          поле
        type: boolean
      description:
        description: включить в ответ описание, необязательное поле
        type: boolean
    type: object
  common.Photo:
    properties:
      id:
        type: integer
      id_advert:
        type: integer
      is_general:
        type: boolean
      url:
        type: string
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:80
info:
  contact: {}
  description: API Server for Advertising Application
  title: Advertising HTTP API
  version: "0.1"
paths:
  /api/advt/:
    get:
      consumes:
      - application/json
      description: getting ordered advert
      operationId: get-by-order
      parameters:
      - description: sort info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/common.AdvertSortOrderParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: GetByOrder
      tags:
      - getByOrder
    post:
      consumes:
      - application/json
      description: create advert
      operationId: create
      parameters:
      - description: advert info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/common.AdvertWithPhoto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Create
      tags:
      - create
  /api/advt/:id/item/:
    get:
      consumes:
      - application/json
      description: get advert
      operationId: get
      parameters:
      - description: field getting info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/common.AdvertFieldParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get
      tags:
      - get
swagger: "2.0"